feat(autoware_euclidean_cluster): porting::autoware_euclidean_cluster::refactor include folder structor, v0.6
