+++ perception/autoware_euclidean_cluster_object_detector/include/autoware/euclidean_cluster_object_detector/euclidean_cluster.hpp
@@ -14,8 +14,8 @@
 
 #pragma once
 
-#include "autoware/euclidean_cluster/euclidean_cluster_interface.hpp"
-#include "autoware/euclidean_cluster/utils.hpp"
+#include "autoware/euclidean_cluster_object_detector/euclidean_cluster_interface.hpp"
+#include "autoware/euclidean_cluster_object_detector/utils.hpp"
 
 #include <pcl/point_types.h>
