+++ perception/autoware_euclidean_cluster_object_detector/lib/euclidean_cluster.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/euclidean_cluster/euclidean_cluster.hpp"
+#include "autoware/euclidean_cluster_object_detector/euclidean_cluster.hpp"
 
 #include <pcl/kdtree/kdtree.h>
 #include <pcl/segmentation/extract_clusters.h>
