+++ perception/autoware_euclidean_cluster_object_detector/src/voxel_grid_based_euclidean_cluster_node.cpp
@@ -14,7 +14,7 @@
 
 #include "voxel_grid_based_euclidean_cluster_node.hpp"
 
-#include "autoware/euclidean_cluster/utils.hpp"
+#include "autoware/euclidean_cluster_object_detector/utils.hpp"
 
 #include <memory>
 #include <vector>
